initialize each carousel type separately (instead of using data-slick)
