+++ events.html
@@ -180,7 +180,7 @@
                 </div>
             </div>
 
-            <section class="carousel u-onlyMobile">
+            <section class="carousel u-onlyMobile js-carousel-standard">
                 <img class="carousel--item carousel--item-cover" src="images/events/2BDD79D7-9C8B-4453-933E-5EDB7F600CBD.JPG">
                 <img class="carousel--item carousel--item-cover" src="images/events/473EC4D2-AE4E-4305-B343-B10198E8C615.JPG">
                 <img class="carousel--item carousel--item-cover" src="images/events/51AB214F-6B94-49DE-A849-1772EDFD7346.JPG">
