+++ index.html
@@ -42,7 +42,7 @@
         </div>
 
         <section class="homepageHeader">
-            <div class="carousel" data-slick='{"autoplay":true,"autoplaySpeed":5000,"fade":true}'>
+            <div class="carousel js-carousel-homepage">
                 <div class="homepageHeader">
                     <img class="homepageHeader--img" src="images/index/city-mural.jpeg">
                     <img class="homepageHeader--svg" src="images/speech-bubble.svg">
@@ -110,7 +110,7 @@
         </section>
 
         <section id="testimonials" class="l-section">
-            <div class="carousel-responsive">
+            <div class="carousel js-carousel-testimonials">
                 <div class="testimonial">
                     <div class="testimonial--pic">
                         <img src="images/index/testimonials/anna-olsen.png">
