+++ programs.html
@@ -163,7 +163,7 @@
                 </div>
             </div>
 
-            <section class="carousel u-onlyMobile">
+            <section class="carousel u-onlyMobile js-carousel-standard">
                 <img class="carousel--item carousel--item-cover" src="images/programs/05CAA216-6086-4FFE-ACC3-A652DE4FFE96.JPG">
                 <img class="carousel--item carousel--item-cover" src="images/programs/1FEAC200-D3B1-4C28-8E12-EE667988F65A.JPG">
                 <img class="carousel--item carousel--item-cover" src="images/programs/2BDD79D7-9C8B-4453-933E-5EDB7F600CBD.JPG">
