+++ scripts/index.js
@@ -1,6 +1,7 @@
 import './vendor/jquery-3.6.0.slim.min.js'
 import './vendor/slick.min.js'
 import './instaFeed.js'
+import './carousel.js'
 
 /*
  * restore animations once dom has settled
@@ -67,32 +68,4 @@ $('#muralGallerySlider').click(event => {
     } else {
         event.target.innerHTML = "&#10095;"
     }
-})
-
-
-/*
- * start carousels
- */
-const slickDefaults = {
-    dots: true,
-    dotsClass: 'carousel--dots',
-    prevArrow: '<button type="button" class="carousel--arrow carousel--arrow-prev">&#10094;</button>',
-    nextArrow: '<button type="button" class="carousel--arrow carousel--arrow-next">&#10095;</button>'
-}
-
-$(document).ready(() => {
-    $('.carousel').slick(slickDefaults)
-})
-
-$(document).ready(() => {
-    $('.carousel-responsive').slick(Object.assign(slickDefaults, {
-        slidesToShow: 2,
-        slidesToScroll: 2,
-        responsive: [
-            {
-                breakpoint: 1000,
-                settings: {slidesToShow: 1, slidesToScroll: 1}
-            }
-        ]
-    }))
 })
