+++ storybook/index.html
@@ -25,7 +25,7 @@
             </nav>
             <main class="l-main">
                 <section class="demo" data-module-name="carousel">
-                    <div class="carousel">
+                    <div class="carousel js-carousel-standard">
                         <img src="https://p.bigstockphoto.com/GeFvQkBbSLaMdpKXF1Zv_bigstock-Aerial-View-Of-Blue-Lakes-And--227291596.jpg">
                         <img src="https://media.istockphoto.com/photos/child-hands-formig-heart-shape-picture-id951945718?k=6&m=951945718&s=612x612&w=0&h=ih-N7RytxrTfhDyvyTQCA5q5xKoJToKSYgdsJ_mHrv0=">
                         <img src="https://cdn.eso.org/images/thumb300y/eso1907a.jpg">
