+++ style/modules/carousel.css
@@ -1,5 +1,4 @@
-.carousel,
-.carousel-responsive {
+.carousel {
     position: relative;
     width: 100%;
 }
